move test specific functions from rc_util.go
